# backend/docs/swagger.yaml
definitions:
...
  diploma_modules_user_handler_model.Address:
    properties:
      description:
        type: string
      id:
        type: integer
      street:
        type: string
    type: object
  diploma_modules_user_handler_model.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  diploma_modules_user_handler_model.User:
    properties:
      address:
        $ref: '#/definitions/diploma_modules_user_handler_model.Address'
      id:
        type: integer
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  model.GetUserProfileResponse:
    properties:
      user:
        $ref: '#/definitions/diploma_modules_user_handler_model.User'
    type: object
...
  model.UpdateUserProfileRequest:
    properties:
      name:
        type: string
      phone_number:
        type: string
    type: object
  model.UserRoleResponse:
    properties:
      role:
        type: integer
    type: object
info:
  contact: {}
  description: This is a sample server with JWT authorization.
  title: Go JWT Swagger Example API
  version: "1.0"
paths:
...
  /api/user/profile:
    get:
      description: Fetch the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetUserProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user profile
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile
      parameters:
      - description: Update Profile Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetUserProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - user
  /api/user/role:
    get:
      description: Returns role ID for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserRoleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/diploma_modules_user_handler_model.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user role
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"